+++ src/components/films-count.js
@@ -1,0 +1,7 @@
+const createFilmsCountTemplate = () => {
+  return (
+    `<section class="footer__statistics">
+      <p>130 291 movies inside</p>
+    </section>`
+  );
+};
